+++ packages/mechanic-shop-fe/src/components/PartList/SingleItemView/index.js
