--- packages/mechanic-shop-fe/src/screens/SingleItemViewScreen.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import SingleItemView from '../components/Inventory/SingleItemView';
-
-const InventoryScreen = () => {
-  return <SingleItemView />;
-};
-
-export default InventoryScreen;
